bench-grpc-python-sklearn-xgb/requirements.txt
grpcio
grpcio-tools
xgboost>=1.3.3
numpy
sklearn